+++ tools/scripts/ManifestBuilder.rb
