+++ tools/scripts/constants.rb
