+++ tools/scripts/forEachLibraryAnimation.rb
