+++ tools/scripts/rebuildDefaultSprites.rb
